Add integration test

+++ clients/src/main/java/org/apache/kafka/clients/admin/ClusterMirrorDesc.java
@@ -86,7 +86,7 @@ public class ClusterMirrorDesc {
         private final long destinationOffset;
         private final long lag;
         private final String state;
-        private final short retryAttempt;
+        private final int retryAttempt;
         private final String errorMessage;
 
         public LeaderStateDesc(TopicPartition topicPartition, long sourceOffset, long destinationOffset,
@@ -120,7 +120,7 @@ public class ClusterMirrorDesc {
             return state;
         }
 
-        public short retryAttempt() {
+        public int retryAttempt() {
             return retryAttempt;
         }
